# pub/openapi.yaml
openapi: 3.0.3
info:
  title: ÚjKréta / KRÁTA Mock API
  description: |
    Diák (ellenorzo) + Tanár (naplo) + Admin API dokumentáció.
    Base: https://nemkapodmeg.onrender.com
  version: 1.1.0
servers:
  - url: https://innensem.onrender.com
    description: Render mock
  - url: http://localhost:8090
    description: Local
...
paths:
...
  /naplo/v3/sajat/TanarAdatlap:
    get:
      tags: [Teacher]
      security: [{ bearerAuth: [] }]
      summary: Tanár adatlap
      responses:
        "200":
          description: Teacher profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
  /naplo/v3/sajat/Tanulok:
    get:
      tags: [Teacher]
      security: [{ bearerAuth: [] }]
      summary: Diákok listája
      responses:
        "200":
          description: Students for teacher
  /naplo/v3/sajat/OsztalyCsoportok:
    get:
      tags: [Teacher]
      security: [{ bearerAuth: [] }]
      summary: Tanár osztályai
      responses:
        "200":
          description: Class groups
  /naplo/v3/sajat/OrarendElemek:
    get:
      tags: [Teacher]
      security: [{ bearerAuth: [] }]
      summary: Tanári órarend
      responses:
        "200":
          description: Lessons
  /naplo/v3/sajat/Ertekelesek:
    get:
      tags: [Teacher]
      security: [{ bearerAuth: [] }]
      summary: Értékelések listája
      responses:
        "200":
          description: Grades
    post:
      tags: [Teacher]
      security: [{ bearerAuth: [] }]
      summary: Új értékelés
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGradeRequest"
      responses:
        "201":
          description: Created grade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Grade"
  /naplo/v3/sajat/HaziFeladatok:
    get:
      tags: [Teacher]
      security: [{ bearerAuth: [] }]
      summary: Házi feladatok
      responses:
        "200":
          description: Homework
    post:
      tags: [Teacher]
      security: [{ bearerAuth: [] }]
      summary: Új házi feladat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHomeworkRequest"
      responses:
        "201":
          description: Created homework
  /naplo/v3/sajat/Mulasztasok:
    get:
      tags: [Teacher]
      security: [{ bearerAuth: [] }]
      summary: Mulasztások
      responses:
        "200":
          description: Omissions
    post:
      tags: [Teacher]
      security: [{ bearerAuth: [] }]
      summary: Új mulasztás
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOmissionRequest"
      responses:
        "201":
          description: Created omission
  /naplo/v3/sajat/BejelentettSzamonkeresek:
    get:
      tags: [Teacher]
      security: [{ bearerAuth: [] }]
      summary: Számonkérések
      responses:
        "200":
          description: Tests
    post:
      tags: [Teacher]
      security: [{ bearerAuth: [] }]
      summary: Új számonkérés
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTestRequest"
      responses:
        "201":
          description: Created test
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    basicAuth:
      type: http
      scheme: basic
  schemas:
...
    NameUid:
      type: object
      properties:
        Uid:
          type: string
        Nev:
          type: string
    NameUidDesc:
      type: object
      properties:
        Uid:
          type: string
        Nev:
          type: string
        Leiras:
          type: string
    Subject:
      type: object
      properties:
        Uid:
          type: string
        Nev:
          type: string
        Kategoria:
          $ref: "#/components/schemas/NameUidDesc"
        SortIndex:
          type: integer
...
    Teacher:
      type: object
      properties:
        Uid:
          type: string
        Nev:
          type: string
        EmailCim:
          type: string
        Telefonszam:
          type: string
        IntezmenyAzonosito:
          type: string
        IntezmenyNev:
          type: string
        OsztalyFonokOsztalyok:
          type: array
          items:
            $ref: "#/components/schemas/NameUid"
        Tantargyak:
          type: array
          items:
            $ref: "#/components/schemas/Subject"
...
    Grade:
      type: object
      properties:
        Uid:
          type: string
        RogzitesDatuma:
          type: string
        Tantargy:
          $ref: "#/components/schemas/Subject"
        Tema:
          type: string
        SzamErtek:
          type: integer
        SzovegesErtek:
          type: string
        ErtekeloTanarNeve:
          type: string
        SulySzazalekErteke:
          type: integer
...
    CreateGradeRequest:
      type: object
      properties:
        TantargyUid:
          type: string
        Tema:
          type: string
        SzamErtek:
          type: integer
        SzovegesErtek:
          type: string
        SulySzazalekErteke:
          type: integer
        Tipus:
          $ref: "#/components/schemas/NameUidDesc"
        OsztalyCsoportUid:
          type: string
        TanuloUid:
          type: string
    CreateHomeworkRequest:
      type: object
      properties:
        TantargyUid:
          type: string
        Szoveg:
          type: string
        Hatarido:
          type: string
        OsztalyCsoportUid:
          type: string
    CreateOmissionRequest:
      type: object
      properties:
        TanuloUid:
          type: string
        Datum:
          type: string
        Tipus:
          $ref: "#/components/schemas/NameUidDesc"
        KesesPercben:
          type: integer
        OsztalyCsoportUid:
          type: string
    CreateTestRequest:
      type: object
      properties:
        TantargyUid:
          type: string
        Datum:
          type: string
        Modja:
          $ref: "#/components/schemas/NameUidDesc"
        OsztalyCsoportUid:
          type: string